requirements.txt
# Specter — Network Pentesting Toolkit
# Instalar con: pip install -r requirements.txt

# Manipulacion de paquetes de red
scapy>=2.5.0

# Wrapper de Nmap
python-nmap>=0.7.1

# SSH y conexiones remotas (post-explotacion)
paramiko>=3.4.0

# Reconocimiento HTTP / OSINT
requests>=2.31.0
dnspython>=2.4.0
shodan>=1.31.0

# Generacion de reportes
jinja2>=3.1.2
markdown>=3.5.1

# Output CLI enriquecido
rich>=13.7.0

# Configuracion por variables de entorno
python-dotenv>=1.0.0

# Utilidades de red
netaddr>=0.10.1
ipwhois>=1.2.0
